refactor(sourceInfo): get body MethodInfo directly from Example
