+++ sln/src/NSpec/Domain/MethodExample.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Reflection;
 
 namespace NSpec.Domain
